convertido valores para decimal para que seja usado corretamente dentro de despesa e receita

+++ Backend/models/schemas.py
@@ -1,13 +1,14 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 
 class Receita(BaseModel):
-    valor: float
+    valor: Decimal
     descricao: str
     categoria: str
 
 
 class Despesa(BaseModel):
-    valor: float
+    valor: Decimal
     descricao: str
     categoria: str
